+++ src/container/container.controller.ts
@@ -1,0 +1,50 @@
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpCode,
+  HttpStatus,
+  Param,
+  ParseIntPipe,
+  Patch,
+  Post,
+} from '@nestjs/common';
+
+import { ContainerDetail, ContainerService } from './container.service';
+import { CreateContainerDto } from './dto/create-container.dto';
+import { UpdateContainerDto } from './dto/update-container.dto';
+
+@Controller('containers')
+export class ContainerController {
+  constructor(private readonly containerService: ContainerService) {}
+
+  @Post()
+  create(@Body() dto: CreateContainerDto) {
+    return this.containerService.create(dto);
+  }
+
+  @Get()
+  findAll() {
+    return this.containerService.findAll();
+  }
+
+  @Get(':id')
+  findOne(@Param('id', ParseIntPipe) id: number): Promise<ContainerDetail> {
+    return this.containerService.findOne(id);
+  }
+
+  @Patch(':id')
+  update(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() dto: UpdateContainerDto,
+  ): Promise<ContainerDetail> {
+    return this.containerService.update(id, dto);
+  }
+
+  @Delete(':id')
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async remove(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    await this.containerService.remove(id);
+  }
+}